+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/testing/WithAdHocTestsLogic.scala
@@ -4,15 +4,19 @@ import io.circe.syntax.EncoderOps
 import io.restassured.RestAssured.`given`
 import io.restassured.module.scala.RestAssuredSupport.AddThenToResponse
 import org.hamcrest.Matchers.equalTo
+import pl.touk.nussknacker.engine.api.graph.ScenarioGraph
+import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.test.NuRestAssureExtensions.{AppConfiguration, EqualsJsonBody, JsonBody}
 import pl.touk.nussknacker.test.base.it.{NuItTest, WithSimplifiedConfigScenarioHelper}
 import pl.touk.nussknacker.test.processes.WithScenarioActivitySpecAsserts.UsersBasicAuth
+import pl.touk.nussknacker.ui.api.description.NodesApiEndpoints.Dtos.TestSourceParameters
 import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.ScenarioTestData
 import pl.touk.nussknacker.ui.api.description.scenarioTesting.Dtos.Validate.ScenarioTestValidationRequest
+import pl.touk.nussknacker.ui.process.marshall.CanonicalProcessConverter
 import pl.touk.nussknacker.ui.process.marshall.CanonicalProcessConverter.toScenarioGraph
 
 trait WithAdHocTestsLogic {
-  self: WithAdHocTestParameters with WithSimplifiedConfigScenarioHelper with NuItTest =>
+  self: WithSimplifiedConfigScenarioHelper with NuItTest =>
 
   def shouldValidateParametersProperly(): Unit = {
     val request = ScenarioTestValidationRequest(
@@ -35,31 +39,6 @@ trait WithAdHocTestsLogic {
            |    "validationErrors": [],
            |    "validationPerformed": true
            |}""".stripMargin
-      )
-  }
-
-  def shouldReturnErrorsForInvalidParameters(): Unit = {
-    val request = ScenarioTestValidationRequest(
-      exampleScenarioGraph,
-      ScenarioTestData.WithParameters(invalidParameters),
-    ).asJson.toString()
-
-    given()
-      .applicationState {
-        createSavedScenario(exampleScenario)
-      }
-      .when()
-      .basicAuthAllPermUser()
-      .jsonBody(request)
-      .post(s"$nuDesignerHttpAddress/api/scenarioTesting/${exampleScenario.name}/validate")
-      .Then()
-      .statusCode(200)
-      .equalsJsonBody(
-        s"""{
-           |    "validationErrors": $expectedValidationErrorsOnInvalidParametersJson,
-           |    "validationPerformed": true
-           |}
-           |""".stripMargin
       )
   }
 
@@ -108,5 +87,17 @@ trait WithAdHocTestsLogic {
        |      "status": "AVAILABLE"
        |    }
        |}""".stripMargin
+
+  protected def exampleScenarioSourceId: String
+
+  protected def exampleScenario: CanonicalProcess
+
+  protected def validParameters: TestSourceParameters
+
+  protected def parametersProvidedForDryRun: String
+
+  protected def expectedTestParametersJson: String
+
+  protected def exampleScenarioGraph: ScenarioGraph = CanonicalProcessConverter.toScenarioGraph(exampleScenario)
 
 }
